moved shader to config
