+++ examples/src/examples/compute/texture-gen/example.mjs
@@ -1,5 +1,6 @@
 import * as pc from 'playcanvas';
 import { deviceType, rootPath } from '@examples/utils';
+import files from '@examples/files';
 
 const canvas = document.getElementById('application-canvas');
 if (!(canvas instanceof HTMLCanvasElement)) {
@@ -90,30 +91,7 @@ assetListLoader.load(() => {
     const shader = new pc.Shader(device, {
         name: 'ComputeShader',
         shaderLanguage: pc.SHADERLANGUAGE_WGSL,
-        cshader: `
-
-            @group(0) @binding(0) var inputTexture: texture_2d<f32>;
-            // @group(0) @binding(1) is a sampler of the inputTexture, but we don't need it in the shader.
-            @group(0) @binding(2) var outputTexture: texture_storage_2d<rgba8unorm, write>;
-
-            // color used to tint the source texture
-            const tintColor: vec4<f32> = vec4<f32>(1.0, 0.7, 0.7, 1.0);
-
-            @compute @workgroup_size(1, 1, 1)
-            fn main(@builtin(global_invocation_id) global_id : vec3u) {
-
-                let uv = vec2i(global_id.xy);
-
-                // load a color from the source texture
-                var texColor = textureLoad(inputTexture, uv, 0);
-
-                // tint it
-                texColor *= tintColor;
-
-                // write it to the output texture
-                textureStore(outputTexture, vec2<i32>(global_id.xy), texColor);
-            }
-        `
+        cshader: files['compute-shader.wgsl']
     });
 
     // bind group for the compute shader - this needs to match the bindings in the shader
